refactor: db/config 파일 server/config에서 관리하도록 수정합니다
db/config/config.js 는 server/config/index.js를 가져오는 역할만 하도록 수정
